[iotedge check] Skip latest version check in nested scenarios (#4461)
Nested deployments typically do not have access to the public internet, and therefore cannot perform the HTTP request to fetch the latest version information from GitHub. This PR addresses this issue by checking if `iotedge check` is running in a nested deployment, and tweaking the behavior of the `aziot-edged-version` check accordingly.

* * *

This PR ended up being a bit gnarlier than expected, as the version-check HTTP request was being made prior to loading the config.yaml, which must be read from to determine whether or not there is a nested deployment. This seems to have been done due to historical reasons - i.e: the asynchronous HTTP request logic was added prior to the refactor that enabled individual checks to spawn async tasks on the tokio runtime. The code has now been moved directly into the relevant check, which makes much more sense from a code-organization standpoint.
